pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "waveos"
version = "0.1.0"
description = "Wave OS demo CLI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.7",
  "rich>=13.7",
  "jinja2>=3.1",
  "prometheus-client>=0.20"
]

[project.scripts]
waveos = "waveos.cli:main"

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "opentelemetry-api>=1.25",
  "opentelemetry-sdk>=1.25",
  "opentelemetry-exporter-otlp>=1.25"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
